Let there be a cloud in the background
The cloud shape was achieved by hit and trial.
I knew it can be done with 4 circles and a rectangle, so trying out a
few combos or arcs with fills in chrome dev tools with canvas open, it
didn’t take long.
Would suggest this exercise as a definite for anyone trying out canvas
for the first time. Really gives a feel for moving things around on
canvas.